mksdk: Don't change ring volume or mode if DND is enabled
Profiles that override the ring volume or mode can disable DND, but
this is not what the user would expect. Ensure that DND never gets
disabled when changing profile and postpone the ring mode and volume
overrides to when DND gets disabled. This allows to give DND an higher
priority and at the same have the overrides of the current profile
always applied.

Change-Id: I065c8df5be924ade63d943414a6da1c29baa3a20